Separated all hard coded const declarations into a single constants file
